flask/requirements.txt
requests~=2.32.3
Flask~=3.1.1
ultralytics~=8.3.152
pillow~=11.2.1